+++ tests/models/test_models.py
@@ -1,4 +1,3 @@
-import copy
 import os
 import pytest
 
@@ -6,11 +5,6 @@ from pathlib import Path
 from pydantic import ValidationError
 
 from BALSAMIC.models.cache_models import ReferenceMeta, ReferenceUrlsModel
-from BALSAMIC.models.metrics_model import (
-    MetricModel,
-    MetricValidationModel,
-    MetricConditionModel,
-)
 from BALSAMIC.models.models import (
     VCFAttributes,
     VarCallerFilter,
@@ -26,161 +20,6 @@ from BALSAMIC.models.models import (
     ParamsVEP,
     AnalysisPonModel,
 )
-
-
-def test_referencemeta():
-    """test ReferenceMeta for correctly building model"""
-    # GIVEN a reference model
-    reference_files = {
-        "basedir": "basedir",
-        "reference_genome": {
-            "url": "gs://some_path/b37/human_g1k_v37.fasta.gz",
-            "file_type": "fasta",
-            "gzip": True,
-            "genome_version": "hg19",
-            "file_name": "genome.fa",
-            "dir_name": "genome",
-        },
-        "dbsnp": {
-            "url": "gs://some_path/b37/dbsnp_138.b37.vcf.gz",
-            "file_type": "fasta",
-            "gzip": True,
-            "genome_version": "hg19",
-            "file_name": "dbsnp.vcf",
-        },
-    }
-
-    # WHEN build the model
-    build_model = ReferenceMeta.parse_obj(reference_files)
-
-    # THEN model should have correct attributes
-    assert build_model.reference_genome.genome_version == "hg19"
-    assert build_model.dbsnp.genome_version == "hg19"
-    assert build_model.reference_genome.get_output_file == "basedir/genome/genome.fa"
-
-
-def test_referenceurlsmodel_build_model():
-    """test ReferenceUrlsModel for correctly building the model"""
-    # GIVEN a reference model
-    dummy_output_file = "some_random_file"
-    dummy_output_path = "some_path"
-    actual_path = Path(dummy_output_path, dummy_output_file).as_posix()
-
-    dummy_reference = {
-        "url": "gs://domain/file_name",
-        "file_type": "fasta",
-        "gzip": True,
-        "genome_version": "hg19",
-        "file_name": dummy_output_file,
-        "dir_name": dummy_output_path,
-    }
-
-    # WHEN building the model
-    built_model = ReferenceUrlsModel.parse_obj(dummy_reference)
-
-    # THEN model should have correct attributes
-    assert built_model.url.scheme == "gs"
-    assert built_model.get_output_file == actual_path
-
-
-def test_referenceurlsmodel_validate_file_type():
-    """test ReferenceUrlsModel for validating file type"""
-    # GIVEN a reference model
-    dummy_output_file = "some_random_file"
-    dummy_output_path = "some_path"
-    actual_path = Path(dummy_output_path, dummy_output_file).as_posix()
-
-    dummy_reference = {
-        "url": "gs://domain/file_name",
-        "file_type": "wrong_type",
-        "gzip": True,
-        "genome_version": "hg19",
-        "file_name": dummy_output_file,
-        "dir_name": dummy_output_path,
-    }
-
-    # WHEN building the model
-
-    # THEN model raise error on validation
-    with pytest.raises(ValidationError) as excinfo:
-        built_model = ReferenceUrlsModel.parse_obj(dummy_reference)
-        assert "not a valid reference file format" in excinfo.value
-
-
-def test_referenceurlsmodel_write_md5(tmp_path_factory):
-    """test ReferenceUrlsModel for writing md5 of the output file"""
-    # GIVEN a reference model
-    dummy_output_file = "some_random_file"
-    dummy_output_path = tmp_path_factory.mktemp("some_path")
-    Path(dummy_output_path, dummy_output_file).write_bytes(os.urandom(8196))
-
-    actual_md5_file = Path(dummy_output_path, dummy_output_file + ".md5")
-
-    dummy_reference = {
-        "url": "gs://domain/file_name",
-        "file_type": "fasta",
-        "gzip": True,
-        "genome_version": "hg19",
-        "file_name": dummy_output_file,
-        "dir_name": dummy_output_path.as_posix(),
-    }
-
-    # WHEN building the model
-    built_model = ReferenceUrlsModel.parse_obj(dummy_reference)
-
-    # THEN when md5 of the file should exist
-    built_model.write_md5
-    assert actual_md5_file.is_file()
-
-
-def test_referenceurlsmodel_write_md5_no_output_file(tmp_path_factory):
-    """test ReferenceUrlsModel for failing to write md5 if outputfile doesn't exist"""
-    # GIVEN a reference model
-    dummy_output_file = "some_random_file"
-    dummy_output_path = tmp_path_factory.mktemp("some_path")
-
-    actual_md5_file = Path(dummy_output_path, dummy_output_file + ".md5")
-
-    dummy_reference = {
-        "url": "gs://domain/file_name",
-        "file_type": "fasta",
-        "gzip": True,
-        "genome_version": "hg19",
-        "file_name": dummy_output_file,
-        "dir_name": dummy_output_path.as_posix(),
-    }
-
-    # WHEN building the model
-    built_model = ReferenceUrlsModel.parse_obj(dummy_reference)
-
-    # THEN when md5 of the file should exist
-    with pytest.raises(FileNotFoundError) as excinfo:
-        built_model.write_md5
-        assert "file does not exist" in excinfo.value
-
-
-def test_referenceurlsmodel_validate_genome_version():
-    """test ReferenceUrlsModel for validating genome version"""
-    # GIVEN a reference model
-    dummy_output_file = "some_random_file"
-    dummy_output_path = "some_path"
-    actual_path = Path(dummy_output_path, dummy_output_file).as_posix()
-
-    dummy_reference = {
-        "url": "gs://domain/file_name",
-        "file_type": "fasta",
-        "gzip": True,
-        "genome_version": "wrong_genome",
-        "file_name": dummy_output_file,
-        "dir_name": dummy_output_path,
-    }
-
-    with pytest.raises(ValidationError) as excinfo:
-        # WHEN building the model
-        built_model = ReferenceUrlsModel.parse_obj(dummy_reference)
-
-        # THEN model raise error on validation
-        assert "not a valid genome version" in excinfo.value
 
 
 def test_vcfattributes():
@@ -425,130 +264,6 @@ def test_params_vep():
 
     # THEN assert values
     assert test_vep_built.vep_filters == "all defaults params"
-
-
-def test_metric_condition_model():
-    """test MetricConditionModel attributes parsing"""
-
-    # GIVEN input attributes
-    metric_condition = {"norm": "gt", "threshold": 1}
-
-    # WHEN building the metric condition model
-    metrics_model = MetricConditionModel(**metric_condition)
-
-    # THEN assert retrieved values from the created model
-    assert metrics_model.dict().items() == metric_condition.items()
-
-
-def test_metric_model_pass_validation():
-    """test MetricModel attributes parsing"""
-
-    # GIVEN input attributes
-    metrics = {
-        "header": None,
-        "id": "ACC1",
-        "input": "ACC1.sorted.mrkdup.hsmetric",
-        "name": "MEDIAN_TARGET_COVERAGE",
-        "step": "multiqc_picard_HsMetrics",
-        "value": 2393.0,
-        "condition": {"norm": "gt", "threshold": 1000.0},
-    }
-
-    # WHEN building the metric model
-    metric_model = MetricModel(**metrics)
-
-    # THEN assert retrieved values from the created model
-    assert metric_model.dict().items() == metrics.items()
-
-
-def test_metric_model_duplication_refactoring():
-    """test MetricModel duplications param refactoring"""
-
-    # GIVEN input attributes
-    metrics = {
-        "header": None,
-        "id": "ACC1",
-        "input": "ACC1_R_1_fastqc.zip",
-        "name": "FastQC_mqc-generalstats-fastqc-percent_duplicates",
-        "step": "multiqc_general_stats",
-        "value": 21.517800000611373,
-        "condition": None,
-    }
-
-    # WHEN building the metric model
-    metric_model = MetricModel(**metrics)
-
-    # THEN assert retrieved values from the created model
-    assert metric_model.name == "PERCENT_DUPLICATION_R1"
-
-
-def test_metric_model_fail_validation():
-    """test MetricModel behaviour for an incorrect input"""
-
-    # GIVEN a non accepted input
-    invalid_input = {"header": None, "id": "ACC1"}
-
-    # THEN the model raises an error due to an incomplete input
-    with pytest.raises(ValueError) as input_exc:
-        MetricModel(**invalid_input)
-    assert f"field required" in str(input_exc.value)
-
-
-def test_metric_validation_model_pass(qc_extracted_metrics):
-    """test MetricValidationModel attribute parsing and positive validation"""
-
-    # WHEN building the MetricValidationModel model
-    model = MetricValidationModel(metrics=qc_extracted_metrics)
-
-    # THEN assert retrieved values from the created model
-    assert model.dict()["metrics"] == qc_extracted_metrics
-
-
-def test_metric_validation_model_fail(qc_extracted_metrics):
-    """test MetricValidationModel for an overly restrictive metric condition"""
-
-    # GIVEN input attributes with a value that does not meet the filtering condition
-    metrics = copy.deepcopy(qc_extracted_metrics)
-    metrics[4]["value"] = 2.0  # GC_DROPOUT set to 2.0 (failing condition)
-
-    # THEN check that the model filters the metric according to its norm
-    with pytest.raises(ValueError) as val_exc:
-        MetricValidationModel(metrics=metrics)
-    assert (
-        f"QC metric {metrics[4]['name']}: {metrics[4]['value']} validation has failed. "
-        f"(Condition: {metrics[4]['condition']['norm']} {metrics[4]['condition']['threshold']}, ID: {metrics[4]['id']})"
-        in str(val_exc.value)
-    )
-
-
-def test_multiple_metric_validation_model_fail(qc_extracted_metrics):
-    """test MetricValidationModel for multiple metrics with failing conditions"""
-
-    # GIVEN input attributes that does not meet the specified conditions
-    metrics = copy.deepcopy(qc_extracted_metrics)
-    metrics[4]["value"] = 2.0  # GC_DROPOUT set to 2.0 (failing condition)
-    metrics[8]["value"] = 0.5  # PCT_TARGET_BASES_500X set to 50% (failing condition)
-
-    # THEN check that the model filters the metrics according to its norm
-    with pytest.raises(ValueError) as val_exc:
-        MetricValidationModel(metrics=metrics)
-    assert "2 validation errors for MetricValidationModel" in str(val_exc.value)
-    assert metrics[4]["name"] in str(val_exc.value)
-    assert metrics[8]["name"] in str(val_exc.value)
-
-
-def test_metric_validation_model_norm_fail(qc_extracted_metrics):
-    """test MetricValidationModel ValueError raising for an operator that it is not accepted"""
-
-    # GIVEN a metric with an incorrect norm attribute
-    metrics = copy.deepcopy(qc_extracted_metrics)
-    metrics[4]["condition"]["norm"] = "lower"
-
-    # THEN model raises an error due to a non accepted norm
-    try:
-        MetricValidationModel(metrics=metrics)
-    except KeyError as key_exc:
-        assert metrics[4]["condition"]["norm"] in str(key_exc)
 
 
 def test_analysis_pon_model(test_data_dir: str):
